Copy cluster proxy from original SNO to upgraded one
This change is part of the cluster reconfiguration, which takes place
after spinning up the seed image onto the target SNO. Cluster
reconfiguration supports a predefined set of the original SNO cluster's
configuration, e.g. IDMS, pull-secret, network configuration.

Specifically, this change addresses fetcing the cluster-wide OCP proxy
configuration of the original SNO and creates the respective JSON manifest to be
applied alongside the rest of the cluster reconfiguration manifests.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>