Update tar usage when saving /etc files
In some cases when there is a large number of /etc files to save,
the list of files gets truncated resulting in an incomplete arhive
which later impacts the upgrade process as installation-configuration
systemd service cannot start.

This change removes the xargs usage and uses tar's files-from option
instead which seems to overcome the issues when having a large number
of files.