Add post-imager completion handling to seedgen orchestrator
Prior to launching the ibu-imager container, the seedgen orchestrator
saves the SeedGenerator CR and associated Secret CR.

Once the ibu-imager container is finished, it restarts kubelet to restore
the operators. When LCA recovers, it recreates the saved SeedGenerator
and Secret CRs from the saved files, and completes the seedgen
orchestration.

The ManagedCluster CR for the cluster is recreated on the hub, if the
orchestrator deleted it prior to seed image generation (ie. if a
hubKubeconfig was provided and the ManagedCluster existed on the hub).

The orchestrator inspects the ibu_imager container to verify it exited
successfully, indicating that the seed image was generated and pushed.

If no errors occurred and the imager was successful, the orchestrator
updates the SeedGenerator CR with conditions to indicate success.
Otherwise, the CR is updated to indicate seed generation failed.

The completed ibu_imager container is not removed from the cluster until
the start of a new seed generation. This allows the user to view the
ibu_imager logs until then, whether successful or failed.

Signed-off-by: Don Penney <dpenney@redhat.com>