precaching: Always pull tagged images
The precache handler checks images to see whether they already exist in
container-storage prior to pulling. In the case of tagged images,
however, this means the image could be stale, as the tag may have been
updated in the registry.

This update modifies the precache handler to never skip pulling tagged
images. If the tag has not been updated, the layers should not get
downloaded, only the manifest, so the pull should be quick.

Signed-off-by: Don Penney <dpenney@redhat.com>