Handle IBU with no rhcos delta
In an IBU where both releases have the same underlying rhcos image, the
parent commit of the deployment has unique commit IDs (due to import
from seed), but the same checksum. When the new deployment is created,
the original parent commit is getting pruned as a result.

In order to avoid this pruning, this update includes the following
changes:
- Add the "--no-prune" option to the "ostree admin deploy" command
- Add a call to "rpm-ostree cleanup -b" to update the base refs

Signed-off-by: Don Penney <dpenney@redhat.com>