Add new `SeedReconfiguration` API
Partially solves MGMT-16483

# Struct organization

## Background

We had a single `ClusterInfo` struct which was overloaded, it was used
for:

* Storing information about the seed cluster
* Storing information about the desired (new SNO) configuration
* For ACM deregistration process, which required the cluster name, this
  entire struct was used

There's no good reason, other than general similarity, for all these use
cases to use the same struct. This was made apparent when we wanted to
have a more formal definition of the desired new SNO configuration,
which will also be used the cluster-relocation-operator which needs to
render such configuration in its relocation ISO.

## Change

This commit changes the current situation by instead having 3 separate
structs:

* `ClusterInfo` is an internal LCA struct (never gets serialized), used
  to store information gathered by `GetClusterInfo` (previously called
  `CreateClusterInfo`)
* `SeedReconfiguration` will store the desired (new SNO) configuration.
  This is the struct that cluster-relocation-operator will render into
  its ISO (IBI) and also the file LCA will create from `ClusterInfo`
  gathered from the original SNO so that the new SNO can have the same
  config as the original cluster
* `SeedClusterInfo` is a struct for storing information about the seed
  cluster. It's stored in the seed image as `manifest.json` and used for
  various purposes during reconfiguration.

## `RecertImagePullSpec` and `SEED_FULL_DOMAIN`

The pull-spec of the recert image was passed from the seed image to the
post-pivot LCA through a systemd unit env file that was rendered from a
template. I removed all this complexity to instead simply include this
pull-spec in `SeedClusterInfo` under `RecertImagePullSpec`

Also removed unused environment variable `SEED_FULL_DOMAIN` that was in
that env file

# Kubeconfig retention

All the certificates and keys needed for kubeconfig retention were
scattered inside files. This commit moves them into proper fields inside
the `SeedReconfiguration` struct

# `SeedFormatVersion` bump

Since this PR breaks the seed format, it bumps `SeedFormatVersion` from
2 to 3