Configurable automatic rollbacks on upgrade failures
This update introduces automatic rollbacks when a failure occurs during
an IBU, as well as adding an init monitor that triggers a rollback
should the configurable timeout expire before the upgrade completes.

An automatic rollback can occur in these cases::
- post-reboot config service-unit failure
- Upgrade stage completion failure
- init-monitor timeout before Upgrade completes

Upon completion of the Upgrade stage, LCA will shutdown and disable the
init-monitor. If the init-monitor times out prior to this, it will
trigger a rollback.

Details on configuration of the automatic rollbacks can be found in
HACKING.md

Signed-off-by: Don Penney <dpenney@redhat.com>