Copy the kubeadmin password from original SNO
Users of OCP expect the default kubeadmin password generated during
installation time to keep working even after an IBU.

To achieve that, LCA will have to copy the password hash from the
kube-system/kubeadmin secret in the original SNO into the
`SeedReconfiguration` struct so it can be set in the equivalent secret
in the new SNO.

Note that not all clusters that we will be upgrading will have such
password, as OCP encourages people, for security reasons, to disable
this password by deleting the secret as outlined here:
https://docs.openshift.com/container-platform/4.14/authentication/remove-kubeadmin.html

In case no such secret exists, LCA will ensure that the secret will
simply be deleted in the new SNO.

The secret hash replacement will be actually done using a new recert
feature (which is activated by simply setting the hash in the recert
config, or setting it to an empty string to signal to recert to delete
the secret). The reason for doing it using recert is that recert is
already modifying etcd secrets while the cluster is offline, and we
don't want to launch a cluster that might still have some old
compromised seed password, even for just a little while.

An alternative solution to doing it in recert would be to have the seed
generator set a dummy password hash that no one in the world knows the
true password of. We arbitrarily chose the recert solution instead.

Note that if the seed cluster for some reason doesn't have the secret,
and the original SNO does have the secret, then it will cause an error,
as recert only supports modifying an existing secret, not creating one
from scratch. We could modify recert to theoretically create the secret,
but there's not much point to do it - we expect all seed clusters to
still have the secret. Another thing worth noting that we can't simply
in this case create the secret after the cluster has already started,
because the password gets rejected by OCP if the creation timestamp of
the secret differs from the kube-system namespace's creation timestamp
by more than 1 hour.