Remove node deletion from seed image creation
This change leaves the seed node object intact, as recert will rename it
to the target SNO's hostname. This resolves an additional etcd rollout,
which is triggered due to an etcd secret that is emptied when the node is
deleted and it can't be adjusted by recert as it's empty.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>