Health check enhancements and associated reconciler changes
This update includes changes to the LCA system health checks and how
they are handled by the Upgrade and SeedGen reconcilers.

The health checks have been updated to remove the existing polling of
background jobs to perform the health checks, replacing this with
performing each check in turn when HealthCheck is called from the
Upgrade and SeedGen reconcilers. Rather than a blocking call, the
reconcilers run the checks and requeue on failure. This provides
multiple benefits:
- If the health checks are continuing to fail post-reboot in an upgrade,
  the user can now trigger a Rollback at any time, rather than waiting
  for the checks to time out. The reconciler is no longer blocked
  waiting for the set of health checks to pass, which means it is able
  to immediately process the Rollback request.
- The IBU and SeedGen CRs are now updated with progress messages to
  indicate a) that the health checks are in-progress and b) which checks
  are failing.

The SeedGen reconciler has been updated to allow for the health checks
to run and requeue on failure. In addition, the progress status
reporting has been improved to provide better information to the user in
case certain steps take more time, rather than simply seeing "In
progress".

Signed-off-by: Don Penney <dpenney@redhat.com>