Updates for review comments and additional test results
This update includes changes based on review comments related to health
check changes and associated enhancements.

In addition, the changes to the reconciler handler flows found issues
due to database interactions. The client.Update function is treated as a
synchonrous call, but the actual database update may not take place
immediately, either due to system delays or connection issues. As a
result, there could be cases where subsequent reconciler loops get a
stale cached IBU CR from the client.Get function, and end up with a
failure updating the CR again (due to updating a stale CR version). In
order to avoid such issues, the following changes are included:
- The beginning of the reconciler loop has been updated to use a
  non-cached client Get call to query the CR. This avoids using a stale
  cached copy of the CR when the API server is not responding, such as in
  a kube-apiserver rollout.
- Avoid an "immediate" requeue after a database update, to allow time
  for the request to be processed. The requeueImmediately function has
  been modified to return a 2 second delay, rather than immediate, to
  avoid future issues. As well, the IBU reconciler has been updated to
  remove the interim requeue that occurred after the stage transition
  validation, instead moving on to handling the stage without requeue.

Signed-off-by: Don Penney <dpenney@redhat.com>