Introduce optional SriovNetworkNodeState health check
If the sriov-network-operator is not ready to process configuration
updates, it can cause upgrade failures when the post-reboot Upgrade
handler attempts to apply extra-manifests with SRIOV config data. While
there is a CSV health check that verifies the operator is installed and
running, it does not ensure the necessary resources have been created.

The SriovNetworkNodeState CR is created by the sriov-network-operator
when its reconcilers first run. If the CR has a Succeeded state, it
indicates the operator should be ready to process the configuration
updates from the extra-manifests.

However, there is a race condition after rebooting to the upgraded
release, where sriov-network-operator can be launched and run its
reconcilers before kube-apiserver is able to service its requests. As a
result, the first time its reconcilers run, it may not see its config
data at all, and does not create the necessary CRs until the next time
the reconciler runs, 5 minutes later.

As a workaround, the upgrade handler checks the results of a failed
health check, adding an annotation to the default SriovNetworkNodePolicy
CR if the failure indicates the SRIOV resources have not yet been
created, kicking the corresponding reconciler to run immediately, rather
than wait for the periodic requeue, which could be minutes later.

Signed-off-by: Don Penney <dpenney@redhat.com>