seedgen: Remove hubKubeconfig and hub interaction
Due to the potential issues stemming from ACM automatically re-importing
the cluster, the support for hubKubeconfig and LCA deletion/restoration
of the ManagedCluster CR had been dropped. The seedgen system validation
now verifies that the cluster is not being actively managed by ACM prior
to starting seed image generation, by checking for the klusterlets CRD.
If present, this indicates ACM is managing the cluster, and the seedgen
request is rejected.

Signed-off-by: Don Penney <dpenney@redhat.com>