Add Rollback CSR Approver service
From the original stateroot point of view, a rollback is effectively
just a recovery from having the node out-of-service for a possibly
extended period of time. Especially when running an IBU within the first
24 hours of deploying a cluster, this means the control plane certificates
for the original release may be expired when the rollback is triggered.

Recovery from expired control plane certificates requires approving
pending CSRs. Until the CSRs are approved and the certificates renewed,
pods do not get scheduled on the cluster.

In order to address this, this update provides a temporary service-unit
that watches for pending CSRs and approves them. The service-unit is
created and enabled in the LCA Upgrade pre-pivot handler on the original
stateroot only, to run during a rollback. If the control plane
certificates expire while the cluster is running the new stateroot, this
service-unit ensures the certificates get renewed to allow the cluster
to recover.  Once the LCA pod is launched, it is then able to complete
the rollback, which now includes removing this temporary service-unit
from the cluster.

Signed-off-by: Don Penney <dpenney@redhat.com>