Change the matching versions of target-ocp-version label
It was required to label the manifests with the full release ocp version
defined in spec.seedImageRef.version.

It's too much effort for user to label the manifests for each z-stream upgrade
, especially when there may be no changes.

Update to match either full release version, X.Y.Z or X.Y

The utility function to load resources from a directory requires sub directories,
update the directory structure of policy manifests to follow that pattern.