IBU: Add rollbackAvailabilityExpiration status field
From the original stateroot point of view, a rollback is effectively
just a recovery from having the node out-of-service for a possibly
extended period of time. Especially when running an IBU within the first
24 hours of deploying a cluster, this means the control plane certificates
for the original release may be expired when the rollback is triggered,
requiring a manual procedure in order to renew them and recover the
cluster.

This update adds a new field to the IBU CR to provide the user with
indication of when a rollback can be performed before the original
release certificates have expired:
.status.rollbackAvailabilityExpiration

Signed-off-by: Don Penney <dpenney@redhat.com>