Revert "Add check for localVolume presence in CleanupStaleBackups"
Reverting commit 094f0d50e13302a4edcea62769ca520353196264 since it is not needed anymore. There was a race in the ensureBackupsDeleted function - fixed in PR#471 - that was provoking OCPBUGS-32939.

Moreover, a check to ignore existing "localvolume" Backup CRs during Prep stage should not be added, given that LCA will try to create a Backup CR with the same name and this will provoke issues during the pre-pivot step.

Signed-off-by: Leonardo Ochoa-Aday <lochoa@redhat.com>