Enforce FIPS compatibility
Solves MGMT-17748

# Background

LCA recently added support for upgrading of FIPS-enabled clusters (see
commit 82460adaf5f3709f707b7f1126f7f180a32ed587).

# Issue

The current implementation does not enforce FIPS compatibility between
the seed image and the cluster being upgraded. This can lead to issues
when the seed image has FIPS enabled but the cluster being upgraded does
not, or vice versa.

# Solution

Add a check for FIPS compatibility between the seed image and the cluster
being upgraded.

# Implementation details

- Add a new field `HasFIPS` to the `SeedClusterInfo` struct to store
  whether the seed image has FIPS enabled or not. This field is set
  based on the value of the `FIPS` field in the `MachineConfig` object
  referenced by the `machineconfiguration.openshift.io/currentConfig`
  label of the only node in the cluster.

- Check for FIPS compatibility when pulling the seed image. If the seed
  image has FIPS enabled but the cluster being upgraded does not, or
  vice versa, return an error.