# Issue / Requirement / Reason for change
As the scale of the project grows, it is important to encourage
documenting changes, while at the same time making it easier for
reviewers to understand the changes being made.

Since each person often works on their own areas of the project, and
since we have contributors from many different locations and timezones,
we don't often get a chance of having technical discussions in real
time, so it's getting harder to keep track and review changes by other
people. We often have to infer the intent of the changes from the code
itself / vague and outdated Jira tickets that were written well before
the implementation.

We also often make changes without much planning or a design document,
making the code the only source of truth, which can sometimes be a bit
cryptic and will make it difficult for future contributors to have
confidence in their changes, as it's hard to understand the intent of
the original code.

# Solution / Feature Overview

Add a structured PR/commit message template that will encourage people
to share information about their changes.

# Implementation Details

The template is a markdown file placed in `.github/pull_request_template.md`
that will be copied into the PR description automatically by GitHub.
Using `git config commit.template .github/pull_request_template.md` (as explained
in the template's checklist) will also configure git to copy the
template into the default user commit message when users use `git
commit` or a Git GUI which supports the `commit.template` feature.

# Other Information

I often find myself scratching my head at PRs, trying to understand the
motivation behind the changes, or having to explicitly ask for it. I
think this template will help me and other reviewers to have more
effective reviews. It will also serve as a form of communication between
the project engineers, and at the same time be a form of archival
documentation of our project decisions and design. Since it's fully
optional, I don't think it will be a burden to anyone who doesn't want
to participate (although I hope they will, as even though it's a bit
annoying having to write it, it's also often annoying having to review
code without it).