Introduce Subscription health check
This update adds a Subscription check to the set of system health
checks, along with optionality for this specific check. The Subscription
check ensures that each subscription CR has no conditions set to true,
which indicates the subscription is healthy.

This check is skipped in the postpivot Upgrade stage handler, as it is
expected that the catalog sources will not have been updated at this
point, and each subscription would therefore be reported as unhealthy.

One implication of the Subscription health check is that there are no
updates pending, meaning that the corresponding operators are running
the latest available version, depending on the subscription
configuration. In order to allow the user to proceed on a cluster
without completing such updates, should they so desire, LCA will look
for an annotation to determine whether to disable or skip this specific
health check. This annotation can be added to either the IBU or Seedgen
CRs to bypass the check altogether, for those steps:

healthchecks.lca.openshift.io/subscriptions: Disabled

Signed-off-by: Don Penney <dpenney@redhat.com>