Add CSV relatedImages to seed containers.list
A CSV may include relatedImages that are not pulled to the seed SNO due
to limited configuration, which can result in the need to pull those
images on the target node when the operator is starting up. To ensure
these images are pulled during precaching step, the relatedImages from
installed CSVs are now included in the seed image containers.list file.

Signed-off-by: Don Penney <dpenney@redhat.com>