IBU: Add automated container storage cleanup to Prep stage
At the start of handling the Prep stage, LCA checks the container
storage disk usage. If it exceeds a specified threshold (50% default),
images will be automatically deleted until disk usage is within the
threshold. The list of images to be deleted is sorted by image Created
timestamp, filtering out in-use or pinned images, with dangling images
deleted first.

The threshold percentage can be overridden by setting an annotation on
the IBU CR. Additionally, this automated cleanup can be disabled
completely with an IBU annotation.

Signed-off-by: Don Penney <dpenney@redhat.com>