Fix client-go dependency issue
# Background / Context

LCA depends on client-go, config-policy-controller and
local-storage-operator

# Issue / Requirement / Reason for change

See #269

There's an old pinning of client.go in
open-cluster-management.io/config-policy-controller that pinned the
version of client-go to `v12.0.0+incompatible` which means we had to pin
our client-go on that version as well.

This makes importing LCA as a library (e.g. in
image-based-install-operator) annoying as they have to copy our replace
directives

config-policy-controller fixed that a while ago on their main branch but
never released a version tag.

# Solution / Feature Overview

Now that they released a tag, we can move to depend on their latest
v0.14.0 instead, which doesn't depend on the weird v12 client-go
version.

# Another issue

Simply trying to move to the latest version of config-policy-controller
doesn't work, because our recently added local-storage-operator
dependency depends on a weird version of client-go called `v1.5.2`.

# Solution for the other issue

This means we have to move our local-storage-operator dependency to the
last version that didn't depend on this weird version (`go get github.com/openshift/local-storage-operator@569393ff42e1b632254b05d1d89f0f1431246cb5`)