The purpose of this PR is to pick up a change to disk cleanup made in rh-ecosystem-edge/preinstall-utils@cf01558
This change is to ensure that in the case of thin provisioning we clean up all necessary disks in the correct order.

This has been tested in assisted-test infra with a 5 node cluster and an artificially created thin provisioning, the cluster installs correctly where it would not have done so before this change.

It was discovered that disk cleanup does not clean disks in every scenario.

When thin provisioned disk mappings are present, the existing installer is not able to clean them up. this is because the `_tdata` and `_tmeta` mappings are the only mappings that will be returned when filtering by associated disk.

The changed library code ensures that we search for a base DM and then order the DMs correctly so that deletion will be performed in the correct order.

Certain customers were running into the scenario above when they have thin volume provisioning.

For more context, please see the commit here

We derive a name for the base DM using the function `getBaseDMDevicesForThinProvisioning` and we ensure proper delete order is adhered to using `sortDMDevicesForDeletion`

This has been tested as part of an assisted-service deployment using assisted-test-infra to create a 5 node cluster. Some thin volumes were created and linked to the installation disk.

These were cleaned correctly when they would not have been cleaned before.

<!---

This is a personal checklist that should be applicable to most PRs. It's good
to go over it in order to make sure you haven't missed anything. If you feel
like some of these points are not relevant to your PR, feel free to keep them
unchecked and if you want also explain why you think they're inapplicable.

- [x] I performed a rough self-review of my changes
- [x] I explained non-trivial motivation for my code using code-comments
- [x] I made sure my code passes linting, tests, and builds correctly
- [x] I have ran the code and made sure it works as intended, and doesn't introduce any obvious regressions
- [x] I have not committed any irrelevant changes (if you did, please point them out and why, ideally separate them into a different PR)
- [x] I added tests (or decided that tests aren't really necessary)
Tests present in the updated library and also test have been performed on assisted-test-infra to ensure that this is working correctly.