MGMT-19505: IBU fails during lca-prep-stateroot-setup: failed to complete stateroot setup
In new version of crio was added a cleanup that if crio doesn't find image mount point it removes container.
When running with chroot and moint image , mount point is not created in the host namespace
and in this case crio removes the image at some point causing preparation logic to fail.
This change fixes the issue by using nsenter command for podman image mount.